requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
python-dotenv
flask
flask-cors
PyJWT
pdfplumber
pymupdf
